requirements.txt
fastapi 
uvicorn 
pypdf 
python-multipart
torch>=2.0.0
transformers>=4.34.0
accelerate>=0.24.0
google-generativeai>=0.3.0